[FEAT]: AppDelegate & SceneDelegate Main에 맞게 수정
